+++ test/test-all.sh
@@ -9,6 +9,7 @@ tests=(
 "test-help"
 "test-quantize"
 "test-reindex"
+"test-degrade"
 "test-sample-face"
 "test-sample-sdiv"
 "test-sample-grid"