+++ test/test-help.sh
@@ -8,7 +8,7 @@ cmpOsLog helpMain
 $CMD sample     > ${TMP}/helpSample.txt	 2>&1
 cmpOsLog helpSample
 
-$CMD degrade    > ${TMP}/helpDegrade.txt	 2>&1
+$CMD degrade     > ${TMP}/helpDegrade.txt 2>&1
 cmpOsLog helpDegrade
 
 $CMD compare   > ${TMP}/helpCompare.txt 2>&1
@@ -19,6 +19,9 @@ cmpOsLog helpReindex
 
 $CMD quantize   > ${TMP}/helpQuantize.txt 2>&1
 cmpOsLog helpQuantize
+
+$CMD dequantize   > ${TMP}/helpDeQuantize.txt 2>&1
+cmpOsLog helpDeQuantize
 
 $CMD sequence   > ${TMP}/helpSequence.txt 2>&1
 cmpOsLog helpSequence